chore: push code related to ABI gen into `noirc_driver`
