+++ tooling/noirc_abi/Cargo.toml
@@ -11,7 +11,6 @@ license.workspace = true
 [dependencies]
 acvm.workspace = true
 iter-extended.workspace = true
-noirc_frontend.workspace = true
 noirc_printable_type.workspace = true
 toml.workspace = true
 serde_json = "1.0"
@@ -25,11 +24,5 @@ strum = "0.24"
 strum_macros = "0.24"
 
 [features]
-bn254 = [
-    "acvm/bn254",
-    "noirc_frontend/bn254",
-]
-bls12_381 = [
-    "acvm/bls12_381",
-    "noirc_frontend/bls12_381",
-]
+bn254 = ["acvm/bn254"]
+bls12_381 = ["acvm/bls12_381"]
